requirements.txt
customtkinter==5.2.2
openai==1.14.3
python-dotenv==1.0.1
pycaw==20230407
AppOpener==1.7
psutil==5.9.8
send2trash==1.8.2
